requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
asttokens==3.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
blockbuster==1.5.25
cbor==1.0.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
cryptography==44.0.3
dataclasses-json==0.6.7
datasets==4.2.0
decorator==5.2.1
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
executing==2.2.1
fastapi==0.119.0
filelock==3.20.0
FlagEmbedding==1.3.5
forbiddenfruit==0.1.4
frozenlist==1.8.0
fsspec==2025.9.0
googleapis-common-protos==1.70.0
grandalf==0.8
grpcio==1.75.1
grpcio-tools==1.75.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
idna==3.11
ijson==3.4.0.post0
importlib_metadata==8.7.0
inscriptis==2.6.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ir_datasets==0.5.11
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.0
langchain-openai==1.0.0
langchain-text-splitters==1.0.0
langgraph==1.0.0
langgraph-api==0.4.43
langgraph-checkpoint==2.1.2
langgraph-cli==0.4.4
langgraph-prebuilt==1.0.0
langgraph-runtime-inmem==0.14.1
langgraph-sdk==0.2.9
langsmith==0.4.37
lxml==6.0.2
lz4==4.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
openai==2.5.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.17.1
pexpect==4.9.0
pillow==12.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyee==11.1.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.3
pyparsing==3.2.5
pyppeteer==2.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.48.0
structlog==25.4.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
trec-car-tools==2.6
truststore==0.10.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unlzw3==0.2.3
urllib3==1.26.20
uvicorn==0.38.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchfiles==1.1.1
wcwidth==0.2.14
websockets==10.4
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zlib-state==0.1.10
zstandard==0.25.0
